Removed hard-coded obseravtory code.
The observatories.dat file is very incomplete and so I've been hardcoding the obs code to 568.  But that is showing up as a problem for precise orbit work.  An update obseravtories.dat is needed.  TODO that we need to fix ephem_earth.c and some other places to correctly parse the extended observatories.dat file that MPC now distributes.