Added a property called r_mag to BKOrbit and made observations a property instead of attribute.
The r_mag value is the average magnitude reported in the MPC line that has been parsed to create the orbit.  If no r band values are available then we use the 'g' and 'i' band values to take a guess at the 'r' band.   Also made the observations attribute a property where the observations are only loaded from the file if they are needed instead of loading at class initialization.