Hacked out the setting of the distance from the observer setting and hard coded to 40 au.
When creating a barycentric coordinate we need the distance to the barycentre.  But I was having some trouble getting the value from Horizons consistently.  So, I've just hard wired the value at 40au.  This value is only important if someone converts the SkyCoord provided into a Barycentric value and thus needs the distance.